requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
streamlit==1.36.0
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
pdfminer.six>=20231228
python-docx>=1.1.0
pydantic>=2.7
sentence-transformers>=3.0.0
spacy>=3.7.4
nltk>=3.8.1